+++ go/docker-run.sh
@@ -7,8 +7,8 @@ function mrcmd_func_go_docker_run() {
   ${MRCORE_TTY_INTERFACE} docker run \
     -it \
     --rm \
-    -v "$(realpath "${GO_LIB_DIR}"):/go" \
-    -v "$(realpath "${APPX_WORK_DIR}"):/opt/app" \
+    -v "$(mrcmd_os_realpath "${GO_LIB_DIR}"):/go" \
+    -v "$(mrcmd_os_realpath "${APPX_WORK_DIR}"):/opt/app" \
     --env "TZ=${APPX_TZ}" \
     --env "APPX_SERVICE_BIND=${GO_WEBAPP_BIND}" \
     --env "APPX_SERVICE_PORT=${GO_WEBAPP_PORT}" \