+++ java/docker-run.sh
@@ -5,7 +5,7 @@ function mrcmd_func_java_docker_run() {
   ${MRCORE_TTY_INTERFACE} docker run \
     -it \
     --rm \
-    -v "$(realpath "${APPX_WORK_DIR}"):/opt/app" \
+    -v "$(mrcmd_os_realpath "${APPX_WORK_DIR}"):/opt/app" \
     --env "TZ=${APPX_TZ}" \
     --env "APPX_SERVICE_BIND=${JAVA_WEBAPP_BIND}" \
     --env "APPX_SERVICE_PORT=${JAVA_WEBAPP_PORT}" \