+++ plantuml/docker-run.sh
@@ -7,7 +7,7 @@ function mrcmd_func_plantuml_docker_run() {
   ${MRCORE_TTY_INTERFACE} docker run \
     -it \
     --rm \
-    -v "$(realpath "${PLANTUML_SOURCE_DIR}"):/data" \
+    -v "$(mrcmd_os_realpath "${PLANTUML_SOURCE_DIR}"):/data" \
     --env "TZ=${APPX_TZ}" \
     "${PLANTUML_DOCKER_IMAGE}" \
     "$@"